Created a Small base for the Support Chat
Created a Small and ugly front in terms of the Chat System
with Support.

Before one can truly claim it's functional, one will need access
to the message entities and the mappings of Users in a deeper
way.

There of - this is just a shell for the looks of it.