Integrated REST Path Interactions through Backend
Integrated so that Users can register into the Database,
along with being able to process a Response.

In light of this, i had to enable CORS interactions, allowing
of credentials and manually interacting with the header
elements to some extent.

Further more, i had to integrate a solution to parsing a
Readable Stream - Which was resolved by .then() functional
parsings.

A user can now register, and the Username they register
with in the Database is Unique.