Implemented Base for Adding Leases to the Database
Added a majority in terms of the functionality of what is to be expected of when one is adding into the Database from the Frontend.

A Fetch call is made, a POST request, and the DB performs some logic in so far as to sully out wether it's a new version or a old one.

Regardless, one can now post in Leases into the Database, albeit some details may need to be finetuned.