Basic Filtering
After many hours of attempting to perform a deeper check in terms of Dates and filtering there of - I have to concede on the matter for now.

I have at least implemented on the basis of being able to filter on Location, Guests, Start Date, End Date and price.

The only problem is that i have not figured out how to compare dates in JPA, as i believe this would require a lot of date formatting to check "ranges" of dates in that sense.

I attempted to resolve this multiple times by splitting up the DB columns into Years, Days and Months and other matters.

None of which gave results that worked as intended.

Alas, i will stick with this version for now.