Integrated so User can Update leasing information
integrated so that User can update their leasing information through a UI that basically checks to verify that its the same username of who is trying to modify the page as the one who is logged in.

This change will be posted to the database and all the versions of the Listing is kept both upon visting into that part - and when updating.

Also fixed a small bug in terms of getters in Store.