Integrated so person can book dates and have their balance deduced
Integrated so the User is actually updated in the DB in accordance to the price they paid for X amount of days - and also integrated checks to prevent a user from being able to book if they have insufficient funds to begin with.

The amount is only deducted if you actually carry through with a booking.