Integrated DTO's
We were specifically asked to implement DTO's for this project.

Whilst in reality we could have achieved the wanted result with merely utilizing @JsonIgnore on the password of the user - which we do here - As there is no real sensetive data in that sense otherwise -

In light of being demanded to provide DTOs in terms of structure, here they are.

Had to do some rework in some small functionality in terms of User being logged in to make this work.